tstttt

+++ ft_strlen.c
@@ -6,16 +6,15 @@
 /*   By: fbelkaid <fbelkaid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/09 22:03:42 by fbelkaid          #+#    #+#             */
-/*   Updated: 2023/12/13 14:40:38 by fbelkaid         ###   ########.fr       */
+/*   Updated: 2023/12/17 10:42:48 by fbelkaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 
-#include <stdio.h>
-
+#include"libft.h"
 size_t ft_strlen(const char *c)
 {
-    int i;
+    size_t i;
     i=0;
     while (c[i])
     {
@@ -23,12 +22,3 @@ size_t ft_strlen(const char *c)
     }
     return i;
 }
-// {
-//     char d[100]="ugsghsgsgf";
-//     size_t j;
-//     j= ft_strlen(d);
-//     size_t v;
-//     v = strlen(d);
-//     printf("%zu\n%zu",j,v);
-
-// }
